Exact potential on GPU using parallel reduction. only for 32x32

+++ mydefs.hpp
@@ -3,7 +3,7 @@
 
 #define NEWLINE printf("\n");
 
-typedef double fptype;
+typedef float fptype;
 typedef unsigned char byte;
 
 #endif // #ifndef  _MYDEFS_H_
